Do not assume the stream to be seekable for parsing. Supports parsing a pipe or network stream.

+++ src/MP4Parser/Boxes/FullBox.cs
@@ -33,7 +33,7 @@ namespace Media.ISO.Boxes
         {
         }
 
-        protected override long HeaderSize => base.HeaderSize + 4;
+        protected override int HeaderSize => base.HeaderSize + 4;
 
         protected override void ParseHeader(BoxReader reader)
         {
